+++ .github/workflows/ci.yml
@@ -20,7 +20,7 @@ jobs:
           - version: '1'
             os: ubuntu-latest
             arch: x86
-          - version: '1.6'
+          - version: '1.9'
             os: ubuntu-latest
             arch: x64
     steps:
