+++ src/linearity.jl
@@ -151,17 +151,17 @@ function detect_new_linearities(rep::Representation, solver; verbose=0, kws...)
     nonlins = _nonlinearity(rep)
     isempty(nonlins) && return Int[]
     model, T = layered_optimizer(solver)
-    verbose >= 2 && (_model = JuMP.direct_model(model))
+    # verbose >= 2 && (_model = JuMP.direct_model(model))
     is_lin = falses(length(nonlins))
     active = trues(length(nonlins))
     # We pass `true` as we break homogeneity of the problem with `sum(λ) == 1`.
     hull = _zero_hull(rep, T)
     λ, cλ, sum_con = _hull(model, T, hull, rep, eachindex(nonlins), true)
-    if verbose >= 2
-        for (i, λ) in enumerate(λ)
-            MOI.set(model, MOI.VariableName(), λ, "λ[$i]")
-        end
-    end
+    # if verbose >= 2
+    #     for (i, λ) in enumerate(λ)
+    #         MOI.set(model, MOI.VariableName(), λ, "λ[$i]")
+    #     end
+    # end
     if !isempty(lins)
         η, _, _ = _hull(model, T, hull, rep, eachindex(lins))
     end
@@ -173,7 +173,7 @@ function detect_new_linearities(rep::Representation, solver; verbose=0, kws...)
     end
     for i in 1:fulldim(rep) # safer than `while ...`
         (count(is_lin) == length(is_lin) || iszero(count(active))) && break
-        verbose >= 2 && println(_model)
+        # verbose >= 2 && println(_model)
         # FIXME stopping when we have enough hyperplanes to prove that it's empty
         #       should also resolve the issue with presolve.
         is_feasible(model, "detecting new linearity (you may need to activate presolve for some solvers, e.g. by replacing `GLPK.Optimizer` with `optimizer_with_attributes(GLPK.Optimizer, \"presolve\" => GLPK.GLP_ON)`).") || break
