+++ src/opt.jl
@@ -9,36 +9,6 @@ MOI.dimension(set::PolyhedraOptSet) = fulldim(set.rep)
 # will be modified so we don't copy it as it would be expensive.
 Base.copy(set::PolyhedraOptSet) = set
 
-struct VariableInSet{V <: JuMP.ScalarVariable, S <: Union{Rep, Projection}} <: JuMP.AbstractVariable
-    variables::Vector{V}
-    set::S
-end
-function JuMP.build_variable(error_fun::Function, variables::Vector{<:JuMP.ScalarVariable}, set::Union{Rep, Projection})
-    if length(variables) != fulldim(set)
-        _error("Number of variables ($(length(variables))) does not match the full dimension of the polyhedron ($(fulldim(set))).")
-    end
-    return VariableInSet(variables, set)
-end
-function JuMP.add_variable(model::JuMP.AbstractModel, v::VariableInSet, names)
-    dim_names = dimension_names(v.set)
-    if dim_names !== nothing
-        names = copy(names)
-        for i in eachindex(names)
-            if isempty(names[i]) && !isempty(dim_names[i])
-                names[i] = dim_names[i]
-            end
-        end
-    end
-    JuMP.add_bridge(model, PolyhedraToLPBridge)
-    JuMP.add_bridge(model, ProjectionBridge)
-    return JuMP.add_variable(model, JuMP.VariablesConstrainedOnCreation(v.variables, _moi_set(v.set)), names)
-end
-_moi_set(set::Rep) = PolyhedraOptSet(set)
-function JuMP.build_constraint(error_fun::Function, func, set::Rep)
-    return JuMP.BridgeableConstraint(
-        JuMP.build_constraint(error_fun, func, _moi_set(set)),
-        PolyhedraToLPBridge)
-end
 
 abstract type AbstractPolyhedraOptimizer{T} <: MOI.AbstractOptimizer end
 
@@ -153,7 +123,6 @@ function layered_optimizer(solver)
     return optimizer, T
 end
 
-_optimize!(model::JuMP.Model) = JuMP.optimize!(model)
 _optimize!(model::MOI.ModelLike) = MOI.optimize!(model)
 
 function _unknown_status(model, status, message)
@@ -195,3 +164,4 @@ function Base.isempty(p::Rep, solver=Polyhedra.linear_objective_solver(p))
     x, cx = MOI.add_constrained_variables(model, PolyhedraOptSet(p))
     return !is_feasible(model, "trying to determine whether the polyhedron is empty.")
 end
+_moi_set(set::Rep) = PolyhedraOptSet(set)
